Added error raising where input land-sea mask is not on the correct grid.  Moved the RegridLandSea plugin into the standardise module, as this is the only context in which it should be called.
